added request user dto and added validation to pom.xml

+++ src/main/java/com/daniyal/bankingsystem/model/User.java
@@ -8,7 +8,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     private String firstName;
 
@@ -29,11 +29,11 @@ public class User {
         this.password = password;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    private void setId(int id) {
+    private void setId(Long id) {
         this.id = id;
     }
